+++ src/Libraries/SmartStore.Data/Migrations/201408071158147_MultistorePoll.cs
@@ -1,0 +1,33 @@
+namespace SmartStore.Data.Migrations
+{
+    using System;
+    using System.Data.Entity.Migrations;
+	using SmartStore.Data.Setup;
+
+	public partial class MultistorePoll : DbMigration, ILocaleResourcesProvider, IDataSeeder<SmartObjectContext>
+    {
+        public override void Up()
+        {
+			AddColumn("dbo.Poll", "LimitedToStores", c => c.Boolean(nullable: false));
+        }
+        
+        public override void Down()
+        {
+			DropColumn("dbo.Poll", "LimitedToStores");
+        }
+
+		public bool RollbackOnFailure
+		{
+			get { return false; }
+		}
+
+		public void Seed(SmartObjectContext context)
+		{
+			context.MigrateLocaleResources(MigrateLocaleResources);
+		}
+
+		public void MigrateLocaleResources(LocaleResourcesBuilder builder)
+		{
+		}
+    }
+}