+++ changelog.md
@@ -19,6 +19,7 @@
 * #90 Directly set order status to completed
 * #413 Orders: Add a PDF export\download of selected orders
 * #69 Award reward points for product reviews
+* #164 Add multistore support for polls
 
 ###Improvements###
 * Task Scheduler: